backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
httpx>=0.27.0

# OpenAI
openai>=1.12.0
openai-whisper>=20230126

# Procesamiento de audio
numpy>=1.24.0
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0

# TTS (versiones compatibles con Python 3.13)
pyttsx3>=2.90
# TTS removido temporalmente por incompatibilidad con Python 3.13
# TTS>=0.22.0

# Base de datos
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Dependencias adicionales necesarias
pywin32>=306; sys_platform == 'win32'
comtypes>=1.2.0
asyncpg>=0.28.0

# Redis
redis>=5.0.0

# Seguridad
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Utilidades
tqdm>=4.66.1
python-dateutil>=2.8.2
python-magic>=0.4.27
loguru>=0.7.2
orjson>=3.9.10

# Desarrollo
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
flake8>=6.1.0
pylint>=3.0.2
pytest-cov>=4.1.0

# Documentación
mkdocs>=1.5.3
mkdocs-material>=9.4.1
mkdocstrings[python]>=0.24.1
